+++ examples/basic/index.js
@@ -5,7 +5,14 @@ const app = new App()
 
 app
   .use(logger())
-  .get('/', (_, res) => void res.send('<h1>Hello World</h1>'))
+  .get(
+    '/',
+    (_, res) =>
+      void res.format({
+        html: () => res.send('<h1>Hello World</h1>'),
+        text: () => res.send('Hello World'),
+      })
+  )
   .get('/page/:page/', (req, res) => {
     res.status(200).send(`
     <h1>Some cool page</h1>
@@ -16,4 +23,4 @@ app
   `)
   })
 
-  .listen(3000)
+  .listen(3000, () => console.log(`Listening on http://localhost:3000`))