+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinyhttp/markdown
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.3.10
+
 ## 0.2.9
 
 ### Patch Changes